Compiles on Xcode 4.5

+++ external/CocosBuilderReader/CCBReader.m
@@ -161,7 +161,7 @@
     }
     else
     {
-        num = current-1;
+        num = (int)current-1;
     }
     
     [self alignBits];
